Replaced Xceed ShortUpDown with custom control.
A new ShortUpDown user control has been created to replace the Xceed WPF Toolkit library. This removes any licensing concerns related to the library and also fixes an annoying issue with Game ID changes not being updated in the bound object. The new control is almost identical in appearance and function, but is much simpler and easier to maintain.